chore: compilerwaarschuwingen opruimen (2026.7.0 / GCC 14.2)
De nieuwe toolchain meldt een paar dingen die er al langer stonden. Geen
ervan brak iets, maar ze verdrinken echte waarschuwingen in ruis.

1. %u met uint32_t (OpenQuattCIC.cpp:72-76, oq_common.yaml:469-475)
   uint32_t is `unsigned long` op ESP32, dus %u klopt formeel niet. Opgelost
   met een expliciete cast naar unsigned -- dezelfde stijl die regel 73 van
   dump_config al gebruikte. Beide types zijn 32-bit, dus verliesvrij.

2. Krappe buffers (oq_supervisory_controlmode.yaml:1699, 1708)
   char[6] past exact op "23:59"+NUL, maar uur/minuut zijn uint8_t, dus de
   compiler rekent met [0,255] en waarschuwt over afkappen. In de praktijk
   gebeurde dat nooit; de buffers zaten wel precies op de rand. Nu 8 en 16
   bytes, zodat een latere wijziging er niet overheen loopt.

3. kconfgen: ignoring malformed line 'COMPILER_OPTIMIZATION_SIZE=y'
   Die regel miste de CONFIG_-prefix en werd dus weggegooid. Prefix toevoegen
   is niet nodig: ESPHome zet CONFIG_COMPILER_OPTIMIZATION_SIZE zelf al, want
   `compiler_optimization` heeft default SIZE (esp32/__init__.py:1667+2507).
   Regel verwijderd -- de -Os-build blijft dus precies gelijk. Te verifiëren
   aan de flash-grootte in CI: die moet rond 1448259 bytes blijven.

+++ components/openquatt_cic/OpenQuattCIC.cpp
@@ -69,11 +69,13 @@ void OpenQuattCIC::loop() {
 void OpenQuattCIC::dump_config() {
   ESP_LOGCONFIG(TAG, "OpenQuatt CIC");
   LOG_UPDATE_INTERVAL(this);
-  ESP_LOGCONFIG(TAG, "  Timeout: %u ms", this->timeout_ms_);
+  // uint32_t is `unsigned long` op ESP32, dus %u vereist een expliciete cast.
+  ESP_LOGCONFIG(TAG, "  Timeout: %u ms", static_cast<unsigned>(this->timeout_ms_));
   ESP_LOGCONFIG(TAG, "  Response buffer size: %u B", static_cast<unsigned>(this->response_buffer_size_));
-  ESP_LOGCONFIG(TAG, "  Backoff start/max: %u / %u ms", this->backoff_start_ms_, this->backoff_max_ms_);
-  ESP_LOGCONFIG(TAG, "  Stale after: %u ms", this->stale_after_ms_);
-  ESP_LOGCONFIG(TAG, "  Feed error trip: %u", this->feed_error_trip_n_);
+  ESP_LOGCONFIG(TAG, "  Backoff start/max: %u / %u ms", static_cast<unsigned>(this->backoff_start_ms_),
+                static_cast<unsigned>(this->backoff_max_ms_));
+  ESP_LOGCONFIG(TAG, "  Stale after: %u ms", static_cast<unsigned>(this->stale_after_ms_));
+  ESP_LOGCONFIG(TAG, "  Feed error trip: %u", static_cast<unsigned>(this->feed_error_trip_n_));
 }
 
 bool OpenQuattCIC::start_fetch_(const std::string &url) {
